Check `PY_310_MIN` instead of `PY_39_MIN`
Python 3.9.1 changed the `type()` of `Literal` instances to be
`_LiteralGenericAlias` instead of `_GenericAlias`.

`is_literal_type()` assumed the change took place in Python 3.9.0 and
checked for it accordingly. Because the assumption was incorrect, it
failed to detect `Literal` types in Python versions [3.9.0, 3.9.1).

This commit changes `is_literal_type()` to check `PY_310_MIN` instead of
`PY_39_MIN` to fix the issue.